Add check and clear to same file
